+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
 
 - Replaces `param_key` with `i18n_key` for attribute lookup in locale file
+- Remove controller and url helpers from backend
+  - No changes required to existing code unless you were using the internal
+    classes directly
 
 ## [1.0.0] - 2022-03-23